pyproject.toml
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
norecursedirs = [
    "*.egg-info",
    ".git",
    ".mypy_cache",
    "node_modules",
    ".pytest_cache",
    ".vscode",
]
python_files = ["test_*.py", "bench_*.py"]
python_functions = ["test_*", "bench_*"]
log_cli = true
log_cli_level = "INFO"
# 默认跳过集成测试，除非显式指定 -m integration
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests as integration tests (deselect with default, run with -m integration)",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alfred"
version = "0.0.1"
description = "Alfred: simple TODO management slack bot"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "APACHE-2.0"}
authors = [{name = "Huang Wei", email = "huangwei@apache.org"}]
dependencies = [
    "aiohttp>=3.8.1",
    "aiosqlite>=0.17.0",
    "APScheduler>=3.9.1",
    "slack-sdk>=3.19.2",
    "slack-bolt>=1.18.1",
    "flask>=2.1.2",
    "croniter>=1.3.8",
    "PyYAML>=6.0",
    "typer>=0.6.1",
    "SQLAlchemy>=2",
    "psycopg>=3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
alfred = "alfred.main:alfred_in"
